Rename Stack based buffer overflow.c to demonstration of Stack based buffer overflow.c
the program takes an input from the user. no matter what the input, the result is the same, because the  if(test_pw())  condition is always true. 
however, when it comes to buffer overflow vulnerability, the attacker can use this to bypass the flow of the program.   
buffer overflow means that the attacker can use input that exceeds the limit defined by the program and as a result impact other registers in the memory:  rbp,rip - stack bottom and instruction pointer.
first 10 characters are user input.
8 characters go to the rbp
the rest of the input goes straight to the instruction pointer .
disassemble the program and wrtite down the adress of the line in the program
that we are trying to access. assume the address is 40h. 
input example: aaaaaaaaaabbbbbbbb40
so in the example the program takes the input and '40' goes to the instruction pointer.

with this action, the attacker plants the address in the ip and bypasses the program's flow