Create microcontroller.asm
~8051 assembly~
UART stands for universal asynchronous receiver / transmitter and defines a protocol, or set of rules, for exchanging serial data between two devices. UART is very simple and only uses two wires between transmitter and receiver to transmit and receive in both directions. 
UART is responsible for sending and receiving a sequence of bits. At the output of a UART these bits are usually represented by logic level voltages.
microcontroller communicates with UART and LED display.