Update microcontroller.asm
~8051 assembly~
in this program, a microcontroller communicates with UART and LED display
UART stands for universal asynchronous receiver / transmitter and defines a protocol, or set of rules, for exchanging serial data between two devices.  
UART is responsible for sending and receiving a sequence of bits. At the output of a UART these bits are usually represented by logic level voltages.
data received by UART is stored in SBUF.  the LED responds to that value.